Added Status box and Responsive Layout

+++ main.py
@@ -10,13 +10,13 @@ import os
 ctk.set_appearance_mode('dark')
 ctk.set_default_color_theme('green')
 
-
-appWidth, appHeight = 600, 300
+appWidth, appHeight = 900, 500
 
 
 class EntryFrame(ctk.CTkFrame):
     def __init__(self, master, conn, data_frame, **kwargs):
         super().__init__(master, **kwargs)
+        # Initialize class instance variables
         self.conn = conn
         self.cursor = self.conn.cursor()
         self.data_frame = data_frame
@@ -26,54 +26,77 @@ class EntryFrame(ctk.CTkFrame):
         self.user_lastname = StringVar()
         self.user_phonenumber = StringVar()
 
+        # Configure Frame Columns
+        self.columnconfigure(0, weight=1)
+        self.columnconfigure(1, weight=1)
+        self.columnconfigure(2, weight=1)
+
+        # Configure Frame Rows
+        self.rowconfigure(6, weight=1)
+
         # Frame Title
         self.entryframe_title = ctk.CTkLabel(self, text='INFORMATION', font=('helvetica', 25, 'bold'),
                                              text_color='white')
-        self.entryframe_title.grid(row=0, column=0, padx=10, pady=(15, 0), columnspan=2, sticky='nsew')
-
-        # Creator
-        self.creator_label = ctk.CTkLabel(self, text='2023 C.Sarmiento', font=('helvetica', 10))
-        self.creator_label.grid(row=6, column=0, padx=10, pady=(5, 2), sticky='sw')
+        self.entryframe_title.grid(row=0, column=0, padx=10, pady=10, columnspan=3, sticky='nsew')
 
         # Frame Icon
         self.man_icon = ctk.CTkImage(Image.open('assets/dark_man.png'), size=(64, 64))
         self.man_icon_label = ctk.CTkLabel(self, image=self.man_icon, text='', anchor='center')
-        self.man_icon_label.grid(row=1, column=0, padx=8, pady=(10, 0), columnspan=2, sticky='ew')
+        self.man_icon_label.grid(row=1, column=0, padx=10, pady=10, columnspan=3, sticky='nsew')
 
         # Frame Labels
-        self.firstname_label = ctk.CTkLabel(self, text='First Name', font=('helvetica', 15), text_color='white',
-                                            anchor='center')
+        self.firstname_label = ctk.CTkLabel(self, text='First Name', font=('helvetica', 15), text_color='white')
         self.firstname_label.grid(row=2, column=0, padx=10, pady=(10, 0), sticky='ew')
 
-        self.lastname_label = ctk.CTkLabel(self, text='Last Name', font=('helvetica', 15), text_color='white',
-                                           anchor='center')
+        self.lastname_label = ctk.CTkLabel(self, text='Last Name', font=('helvetica', 15), text_color='white')
         self.lastname_label.grid(row=3, column=0, padx=10, pady=(10, 0), sticky='ew')
 
-        self.phone_number_label = ctk.CTkLabel(self, text='Phone Number', font=('helvetica', 15), text_color='white',
-                                               anchor='center')
+        self.phone_number_label = ctk.CTkLabel(self, text='Phone Number', font=('helvetica', 15), text_color='white', )
         self.phone_number_label.grid(row=4, column=0, padx=10, pady=(10, 0), sticky='ew')
 
         # Frame Entries
         self.firstname_entry = ctk.CTkEntry(self, placeholder_text='Enter First Name', textvariable=self.user_firstname)
-        self.firstname_entry.grid(row=2, column=1, padx=10, pady=(10, 0), sticky='ew')
+        self.firstname_entry.grid(row=2, column=1, padx=10, pady=(10, 0), columnspan=2, sticky='ew')
 
         self.lastname_entry = ctk.CTkEntry(self, placeholder_text='Enter Last Name', textvariable=self.user_lastname)
-        self.lastname_entry.grid(row=3, column=1, padx=10, pady=(10, 0), sticky='ew')
+        self.lastname_entry.grid(row=3, column=1, padx=10, pady=(10, 0), columnspan=2, sticky='ew')
 
         self.phonenumber_entry = ctk.CTkEntry(self, placeholder_text='Enter Phone Number',
-                                              textvariable=self.user_phonenumber,)
-        self.phonenumber_entry.grid(row=4, column=1, padx=10, pady=(10, 0), sticky='ew')
+                                              textvariable=self.user_phonenumber, )
+        self.phonenumber_entry.grid(row=4, column=1, padx=10, pady=(10, 0), columnspan=2, sticky='ew')
 
         # Frame Buttons
-        self.add_button = ctk.CTkButton(self, text='Add', cursor='hand2', command=self.add_contact, width=80)
-        self.add_button.grid(row=5, column=0, padx=10, pady=(15, 0), sticky='w')
+        self.add_button = ctk.CTkButton(self, text='Add', cursor='hand2', command=self.add_contact)
+        self.add_button.grid(row=5, column=0, padx=10, pady=(10, 0), sticky='ew')
         self.add_button.bind('<Return>', self.add_contact)
 
-        self.update_button = ctk.CTkButton(self, text='Update', cursor='hand2', command=self.update_contact, width=80)
-        self.update_button.grid(row=5, column=0, columnspan=2, padx=10, pady=(15, 0))
+        self.update_button = ctk.CTkButton(self, text='Update', cursor='hand2', command=self.update_contact)
+        self.update_button.grid(row=5, column=1, padx=10, pady=(10, 0), sticky='ew')
 
-        self.delete_button = ctk.CTkButton(self, text='Delete', cursor='hand2', command=self.delete_contact, width=80)
-        self.delete_button.grid(row=5, column=1, padx=10, pady=(15, 0), sticky='e')
+        self.delete_button = ctk.CTkButton(self, text='Delete', cursor='hand2', command=self.delete_contact)
+        self.delete_button.grid(row=5, column=2, padx=10, pady=(10, 0), sticky='ew')
+
+        self.clear_statusbox_button = ctk.CTkButton(self, text='Clear', cursor='hand2', command=self.clear_statusbox)
+        self.clear_statusbox_button.grid(row=7, column=2, padx=10, pady=10, sticky='ew')
+
+        # Status Text Box
+        self.statusbox = ctk.CTkTextbox(self, corner_radius=10, fg_color='#040D12', font=('consolas', 12), text_color='white')
+        self.statusbox.grid(row=6, column=0, padx=10, pady=(10, 0), columnspan=3, sticky='nsew')
+
+        # Frame Mark
+        self.framemark_label = ctk.CTkLabel(self, text='C.Sarmiento 2023 Version 1.1', font=('helvetica', 10))
+        self.framemark_label.grid(row=7, column=0, padx=10, pady=(0, 2), sticky='sw')
+
+    def add_status(self, text):
+        status = f"$ {text}\n"
+        self.statusbox.configure(state='normal')
+        self.statusbox.insert("end", status)
+        self.statusbox.configure(state='disable')
+
+    def clear_statusbox(self):
+        self.statusbox.configure(state='normal')
+        self.statusbox.delete('0.0', 'end')
+        self.statusbox.configure(state='disable')
 
     def get_entry_data(self):
         return self.user_firstname.get().strip().title(), self.user_lastname.get().strip().title(), self.user_phonenumber.get().lstrip(
@@ -101,8 +124,7 @@ class EntryFrame(ctk.CTkFrame):
                     self.cursor.execute(query, datacontact)
                     self.data_frame.clear()
                     self.conn.commit()
-                    CTkMessagebox(title='Add Contact', message='Contact Added Successfully.', icon='check',
-                                  option_1='Thanks')
+                    self.add_status(f'New Contact Added Successfully.')
                 else:
                     CTkMessagebox(title='Contact Record', message='Phone number already in Record.')
                 self.clear_entrybox()
@@ -114,69 +136,93 @@ class EntryFrame(ctk.CTkFrame):
 
     def update_contact(self):
         user_firstname, user_lastname, user_phonenumber = self.get_entry_data()
-        response = CTkMessagebox(title='Confirm Update', message='Are you sure to update this contact?',
-                                 icon='question', option_1="No", option_2="Yes")
-        if response.get() == 'Yes':
-            query = 'UPDATE contacts SET FirstName = ?, LastName = ? WHERE PhoneNumber = ?'
-            datacontact = (user_firstname, user_lastname, user_phonenumber)
-            self.cursor.execute(query, datacontact)
-            self.conn.commit()
-            self.data_frame.clear()
-            self.clear_entrybox()
+        if user_firstname and user_lastname and user_phonenumber != '':
+            response = CTkMessagebox(title='Confirm Update', message='Are you sure to update this contact?',
+                                     icon='question', option_1="No", option_2="Yes")
+            if response.get() == 'Yes':
+                query = 'UPDATE contacts SET FirstName = ?, LastName = ? WHERE PhoneNumber = ?'
+                datacontact = (user_firstname, user_lastname, user_phonenumber)
+                self.cursor.execute(query, datacontact)
+                self.conn.commit()
+                self.data_frame.clear()
+                self.clear_entrybox()
+                self.add_status('Contact Updated Successfully.')
+            else:
+                pass
         else:
-            pass
+            CTkMessagebox(title='Warning', message='No Contact selected to update.', icon='warning', sound=True)
 
     def delete_contact(self):
         user_firstname, user_lastname, user_phonenumber = self.get_entry_data()
-        response = CTkMessagebox(title='Confirm Delete', message='Are you sure to delete this contact?',
-                                 icon='question',
-                                 option_1="No", option_2="Yes")
-        if response.get() == 'Yes':
-            query = 'DELETE FROM contacts WHERE FirstName = ? AND LastName = ? AND PhoneNumber = ?'
-            datacontact = (user_firstname, user_lastname, user_phonenumber)
-            self.cursor.execute(query, datacontact)
-            self.conn.commit()
-            self.data_frame.clear()
-            self.clear_entrybox()
+        if user_firstname and user_lastname and user_phonenumber != '':
+            response = CTkMessagebox(title='Confirm Delete', message='Are you sure to delete this contact?',
+                                     icon='question',
+                                     option_1="No", option_2="Yes")
+            if response.get() == 'Yes':
+                query = 'DELETE FROM contacts WHERE FirstName = ? AND LastName = ? AND PhoneNumber = ?'
+                datacontact = (user_firstname, user_lastname, user_phonenumber)
+                self.cursor.execute(query, datacontact)
+                self.conn.commit()
+                self.data_frame.clear()
+                self.clear_entrybox()
+                self.add_status('Contact Deleted Successfully.')
+            else:
+                pass
         else:
-            pass
+            CTkMessagebox(title='Warning', message='No Contact selected to delete.', icon='warning', sound=True)
 
 
 class DataFrame(ctk.CTkFrame):
     def __init__(self, master, conn, entry_frame, **kwargs):
         super().__init__(master, **kwargs)
+        # Initialize class instance variables
         self.conn = conn
         self.cursor = self.conn.cursor()
         self.entry_frame = entry_frame
 
-        # # Define StringVar
+        # Define StringVar
         self.to_search = StringVar()
         self.to_search.trace('w', self.search)
 
+        # Configure Frame Columns
+        self.columnconfigure(0, weight=1)
+        self.columnconfigure(1, weight=1)
+
+        # Configure Frame Rows
+        self.rowconfigure(2, weight=1)
+
         # Frame Title
         self.dataframe_title = ctk.CTkLabel(self, text='RECORD', font=('helvetica', 25, 'bold'), text_color='white')
-        self.dataframe_title.grid(row=0, column=0, padx=10, pady=(15, 0), columnspan=3, sticky='ew')
+        self.dataframe_title.grid(row=0, column=0, padx=10, pady=10, columnspan=3, sticky='nsew')
 
         # Label
         self.search_label = ctk.CTkLabel(self, text='Search:', font=('helvetica', 15), text_color='white')
-        self.search_label.grid(row=1, column=0, padx=(10, 0), pady=(10, 0), sticky='w')
+        self.search_label.grid(row=1, column=0, padx=10, pady=10, sticky='e')
 
         # Frame Entries
-        self.to_search_entry = ctk.CTkEntry(self, placeholder_text='Search', textvariable=self.to_search, width=190)
-        self.to_search_entry.grid(row=1, column=0, columnspan=2, padx=(10, 0), pady=(10, 0), sticky='e')
+        self.to_search_entry = ctk.CTkEntry(self, placeholder_text='Search', textvariable=self.to_search)
+        self.to_search_entry.grid(row=1, column=1, padx=10, pady=10, sticky='ew')
 
         # Button
         self.clear_button = ctk.CTkButton(self, text='Clear', cursor='hand2', command=self.clear)
-        self.clear_button.grid(row=1, column=2, padx=10, pady=(10, 0), sticky='ew')
+        self.clear_button.grid(row=1, column=2, padx=10, pady=10, sticky='ew')
+
+        # Treeview Customisation (based on CTk Theme)
+        # self.bg_color = self._apply_appearance_mode(ctk.ThemeManager.theme['CTkFrame']['fg_color'])
+        # self.text_color = self._apply_appearance_mode(ctk.ThemeManager.theme['CTkLabel']['text_color'])
+        # self.selected_color = self._apply_appearance_mode(ctk.ThemeManager.theme['CTkButton']['fg_color'])
 
         # Configure Treeview Style
         self.style = ttk.Style()
         self.style.theme_use('clam')
         self.style.configure("Treeview.Heading", font=('helvetica', 15, 'bold'))
-        self.style.configure("Treeview", rowheight=30, font=('Calibri', 15))
+        self.style.configure("Treeview", rowheight=30, font=('times new romans', 15), borderwidth=5)
+        # self.style.configure("Treeview", rowheight=30, font=('Calibri', 15), background=self.bg_color,
+        #                      foreground=self.text_color, fieldbackground=self.bg_color, borderwidth=5)
+        # self.style.map('Treeview', background=[('selected', self.bg_color)], foreground=[('selected', self.selected_color)])
 
         # Create Treeview
-        self.db_view = ttk.Treeview(self, columns=("firstname", "lastname", "phonenumber"), show='headings', height=9)
+        self.db_view = ttk.Treeview(self, columns=("firstname", "lastname", "phonenumber"), show='headings')
 
         # Place Scrollbar
         self.scrollbar = ctk.CTkScrollbar(self, orientation='vertical')
@@ -194,6 +240,7 @@ class DataFrame(ctk.CTkFrame):
         # Bind Double Left Click
         self.db_view.bind('<Double 1>', self.getrow)
 
+        # Fetch all data from contacts and display on DataFrame
         self.cursor.execute('SELECT * FROM contacts')
         self.rows = self.cursor.fetchall()
         self.update_dataframe(self.rows)
@@ -231,19 +278,50 @@ class DataFrame(ctk.CTkFrame):
 class App(ctk.CTk):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
+        # Get Access Database Directory
         current_directory = os.getcwd()
         db_file_path = f'{current_directory}\\Phonebook.accdb'
 
-        config = (
+        # Set Connection Configurations
+        self.config = (
             r'DRIVER={Microsoft Access Driver (*.mdb, *.accdb)};'
             f'DBQ={db_file_path}'
         )
-        self.conn = pyodbc.connect(config)
+
+        # Create connection
+        self.conn = pyodbc.connect(self.config)
 
         # Create a cursor to execute SQL commands
         self.cursor = self.conn.cursor()
 
         table_exists = False
+
+        # Create a Window
+        self.title('Phonebook')
+        self.iconbitmap('assets/telephone-directory.ico')
+        self.minsize(appWidth, appHeight)
+        self.configure(fg_color='#040D12')
+
+        # Configure Window Column
+        self.columnconfigure(0, weight=1)
+        self.columnconfigure(1, weight=1)
+
+        # Configure Window Row
+        self.rowconfigure(0, weight=1)
+
+        # Initialize EntryFrame
+        self.entry_frame = EntryFrame(master=self, conn=self.conn, data_frame=None, fg_color='#183D3D',
+                                      border_color='#AEC3AE', border_width=2,
+                                      width=300, height=300)
+        self.entry_frame.grid(row=0, column=0, padx=10, pady=10, sticky='nsew')
+
+        # Initialize DataFrame
+        self.data_frame = DataFrame(master=self, conn=self.conn, entry_frame=self.entry_frame, fg_color='#5C8374',
+                                    border_color='white', border_width=2,
+                                    width=300, height=300)
+        self.data_frame.grid(row=0, column=1, padx=(0, 10), pady=10, sticky='nsew')
+
+        self.entry_frame.data_frame = self.data_frame
 
         try:
             # Attempt to fetch data from the 'Contacts' table
@@ -252,7 +330,7 @@ class App(ctk.CTk):
             table_exists = True
         except pyodbc.Error as e:
             # Handle the exception (e.g., print a message)
-            print(f"Table 'Contacts' does not exist or cannot be accessed: {e}")
+            self.entry_frame.add_status(f"Table 'Contacts' does not exist or cannot be accessed: {e}")
 
         if not table_exists:
             # Create the 'Contacts' table if it doesn't exist
@@ -265,46 +343,13 @@ class App(ctk.CTk):
             """
             self.cursor.execute(create_table_sql)
             self.conn.commit()
-            print("Table 'Contacts' created.")
+            self.entry_frame.add_status("Table 'Contacts' created.")
         else:
-            print("Table 'Contacts' already exists.")
-
-        # Create Window
-        self.title('Phonebook')
-        self.iconbitmap('assets/telephone-directory.ico')
-        # self.geometry(f"{appWidth}x{appHeight}")
-        self.minsize(appWidth, appHeight)
-        # self.resizable(False, False)
-        self.configure(fg_color='#040D12')
-        self.columnconfigure(0, weight=1)
-        self.columnconfigure(1, weight=1)
-        self.rowconfigure(0, weight=1)
-
-        self.entry_frame = EntryFrame(master=self, conn=self.conn, data_frame=None, fg_color='#183D3D',
-                                      border_color='#AEC3AE', border_width=2,
-                                      width=300, height=300)
-
-        self.entry_frame.grid(row=0, column=0, padx=10, pady=10, sticky='nsew')
-        self.entry_frame.columnconfigure(0, weight=1)
-        self.entry_frame.columnconfigure(1, weight=1)
-
-        self.data_frame = DataFrame(master=self, conn=self.conn, entry_frame=self.entry_frame, fg_color='#5C8374',
-                                    border_color='white',
-                                    border_width=2, width=300, height=300)
-        self.data_frame.grid(row=0, column=1, padx=(0, 10), pady=10, sticky='nsew')
-        self.data_frame.columnconfigure(0, weight=1)
-        self.data_frame.columnconfigure(1, weight=1)
-        self.data_frame.rowconfigure(2, weight=1)
-
-        self.entry_frame.data_frame = self.data_frame
-
-
-def main():
-    app = App()
-    app.mainloop()
-    app.conn.close()
-    app.cursor.close()
+            self.entry_frame.add_status("Table 'Contacts' accessed.")
 
 
 if __name__ == '__main__':
-    main()
+    app = App()
+    app.mainloop()
+    app.cursor.close()
+    app.conn.close()
